requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.20
Werkzeug==2.3.7
WTForms==3.0.1

# Authentication & Security
email-validator==2.0.0
PyJWT==2.8.0
cryptography==41.0.4

# Image Processing
Pillow==10.0.0

# PDF Generation
reportlab==4.0.5

# OCR & ML
easyocr==1.7.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy>=1.9.0

# Deep Learning (Optional - for advanced features)
 tensorflow>=2.10.0

# PDF Processing
PyPDF2==3.0.1

# Background Tasks
APScheduler==3.10.1

# WebSocket
python-socketio==5.9.0
eventlet==0.33.3

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
python-engineio==4.8.1
# Development
Werkzeug==2.3.7